+++ packages/tswebextension/CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.6] - 2026-03-26
+
+### Changed
+
+- Updated [@adguard/agtree] to `v4.0.4`.
+- Updated [@adguard/tsurlfilter] to `v4.0.5`.
+
+[4.0.6]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v4.0.6
+
+## [4.0.5] - 2026-03-24
+
+### Changed
+
+- Updated [@adguard/agtree] to `v4.0.3`.
+- Updated [@adguard/tsurlfilter] to `v4.0.4`.
+- Updated [@adguard/scriptlets] to `v2.3.1`.
+
+[4.0.5]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tswebextension-v4.0.5
+
 ## [4.0.4] - 2026-03-18
 
 ### Changed